agentic-rag-chat/pyproject.toml
[project]
name = "agentic-rag-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docling>=2.66.0",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-core>=1.2.6",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.6",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
